+++ apps/remnant2toolkit/app/(builds)/_features/filters/_components/build-filters.tsx
@@ -19,29 +19,30 @@ import { InputWithClear } from '@/app/_components/input-with-clear';
 import { EXCLUDE_ITEM_SYMBOL } from '@/app/_constants/item-symbols';
 import { QualityBuildDialog } from '@/app/(builds)/_components/quality-build-dialog';
 import { DEFAULT_BUILD_FIELDS } from '@/app/(builds)/_features/filters/_constants/default-build-fields';
-import { amuletFilter } from '@/app/(builds)/_features/filters/_libs/amulet-filter';
-import { archetypeFilter } from '@/app/(builds)/_features/filters/_libs/archetype-filter';
-import { archetypeSlotFilter } from '@/app/(builds)/_features/filters/_libs/archetype-slot-filter';
-import { buildTagFilter } from '@/app/(builds)/_features/filters/_libs/build-tag-filter';
-import { fusionFilter } from '@/app/(builds)/_features/filters/_libs/fusion-filter';
-import { handGunFilter } from '@/app/(builds)/_features/filters/_libs/hand-gun-filter';
-import { legendaryFragmentFilter } from '@/app/(builds)/_features/filters/_libs/legendary-fragment-filter';
-import { longGunFilter } from '@/app/(builds)/_features/filters/_libs/long-gun-filter';
-import { meleeFilter } from '@/app/(builds)/_features/filters/_libs/melee-filter';
-import { modFilter } from '@/app/(builds)/_features/filters/_libs/mod-filter';
-import { mutatorFilter } from '@/app/(builds)/_features/filters/_libs/mutator-filter';
+import { amuletFilter } from '@/app/(builds)/_features/filters/_libs/filters/amulet-filter';
+import { archetypeFilter } from '@/app/(builds)/_features/filters/_libs/filters/archetype-filter';
+import { archetypeSlotFilter } from '@/app/(builds)/_features/filters/_libs/filters/archetype-slot-filter';
+import { buildTagFilter } from '@/app/(builds)/_features/filters/_libs/filters/build-tag-filter';
+import { fusionFilter } from '@/app/(builds)/_features/filters/_libs/filters/fusion-filter';
+import { handGunFilter } from '@/app/(builds)/_features/filters/_libs/filters/hand-gun-filter';
+import { legendaryFragmentFilter } from '@/app/(builds)/_features/filters/_libs/filters/legendary-fragment-filter';
+import { longGunFilter } from '@/app/(builds)/_features/filters/_libs/filters/long-gun-filter';
+import { meleeFilter } from '@/app/(builds)/_features/filters/_libs/filters/melee-filter';
+import { modFilter } from '@/app/(builds)/_features/filters/_libs/filters/mod-filter';
+import { mutatorFilter } from '@/app/(builds)/_features/filters/_libs/filters/mutator-filter';
+import { relicFilter } from '@/app/(builds)/_features/filters/_libs/filters/relic-filter';
+import { relicFragmentFilter } from '@/app/(builds)/_features/filters/_libs/filters/relic-fragment-filter';
+import { ringFilter } from '@/app/(builds)/_features/filters/_libs/filters/ring-filter';
+import { searchTextFilter } from '@/app/(builds)/_features/filters/_libs/filters/search-text-filter';
+import { skillFilter } from '@/app/(builds)/_features/filters/_libs/filters/skill-filter';
+import { traitFilter } from '@/app/(builds)/_features/filters/_libs/filters/trait-filter';
+import { withCollectionFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-collection';
+import { withOptionalPrismFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-optional-prism';
+import { withPatchAffectedFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-patch-affected-filter';
+import { withQualityFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-quality-filter';
+import { withReferenceFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-reference-filter';
+import { withVideoFilter } from '@/app/(builds)/_features/filters/_libs/filters/with-video-filter';
 import { parseUrlParams } from '@/app/(builds)/_features/filters/_libs/parse-url-params';
-import { relicFilter } from '@/app/(builds)/_features/filters/_libs/relic-filter';
-import { relicFragmentFilter } from '@/app/(builds)/_features/filters/_libs/relic-fragment-filter';
-import { ringFilter } from '@/app/(builds)/_features/filters/_libs/ring-filter';
-import { searchTextFilter } from '@/app/(builds)/_features/filters/_libs/search-text-filter';
-import { skillFilter } from '@/app/(builds)/_features/filters/_libs/skill-filter';
-import { traitFilter } from '@/app/(builds)/_features/filters/_libs/trait-filter';
-import { withCollectionFilter } from '@/app/(builds)/_features/filters/_libs/with-collection';
-import { withPatchAffectedFilter } from '@/app/(builds)/_features/filters/_libs/with-patch-affected-filter';
-import { withQualityFilter } from '@/app/(builds)/_features/filters/_libs/with-quality-filter';
-import { withReferenceFilter } from '@/app/(builds)/_features/filters/_libs/with-reference-filter';
-import { withVideoFilter } from '@/app/(builds)/_features/filters/_libs/with-video-filter';
 import { type BuildFilterFields } from '@/app/(builds)/_features/filters/_types/build-filter-fields';
 
 export function BuildFilters({
@@ -393,6 +394,15 @@ export function BuildFilters({
       params.set(
         withCollectionFilter.buildFilterKey,
         newFilters.withCollection.toString(),
+      );
+    }
+
+    if (newFilters.withOptionalPrism === defaultFilters.withOptionalPrism) {
+      params.delete(withOptionalPrismFilter.buildFilterKey);
+    } else {
+      params.set(
+        withOptionalPrismFilter.buildFilterKey,
+        newFilters.withOptionalPrism.toString(),
       );
     }
 
@@ -798,14 +808,6 @@ export function BuildFilters({
                   <BaseListbox
                     className="mt-1"
                     value={unappliedFilters.withCollection}
-                    onBlur={() => {
-                      if (
-                        unappliedFilters.withCollection !==
-                        filters.withCollection
-                      ) {
-                        applyUrlFilters(unappliedFilters);
-                      }
-                    }}
                     onChange={(value) => {
                       const newFilters = {
                         ...unappliedFilters,
@@ -862,13 +864,6 @@ export function BuildFilters({
                   <BaseListbox
                     className="mt-1"
                     value={unappliedFilters.withQuality}
-                    onBlur={() => {
-                      if (
-                        unappliedFilters.withQuality !== filters.withQuality
-                      ) {
-                        applyUrlFilters(unappliedFilters);
-                      }
-                    }}
                     onChange={(value) => {
                       const newFilters = {
                         ...unappliedFilters,
@@ -891,6 +886,31 @@ export function BuildFilters({
                   </BaseButton>
                 </BaseField>
                 <BaseField
+                  id="with-optional-prism-filter"
+                  className="col-span-full sm:col-span-1"
+                >
+                  <BaseLabel className="text-surface-solid h-[40px] !text-sm font-medium">
+                    {withOptionalPrismFilter.label}
+                  </BaseLabel>
+                  <BaseListbox
+                    className="mt-1"
+                    value={unappliedFilters.withOptionalPrism}
+                    onChange={(value) => {
+                      const newFilters = {
+                        ...unappliedFilters,
+                        withOptionalPrism: value,
+                      };
+                      setUnappliedFilters(newFilters);
+                    }}
+                  >
+                    {withOptionalPrismFilter.options.map(({ label, value }) => (
+                      <BaseListboxOption key={label} value={value}>
+                        <BaseListboxLabel>{label}</BaseListboxLabel>
+                      </BaseListboxOption>
+                    ))}
+                  </BaseListbox>
+                </BaseField>
+                <BaseField
                   id="with-video-filter"
                   className="col-span-full sm:col-span-1"
                 >
@@ -900,11 +920,6 @@ export function BuildFilters({
                   <BaseListbox
                     className="mt-1"
                     value={unappliedFilters.withVideo}
-                    onBlur={() => {
-                      if (unappliedFilters.withVideo !== filters.withVideo) {
-                        applyUrlFilters(unappliedFilters);
-                      }
-                    }}
                     onChange={(value) => {
                       const newFilters = {
                         ...unappliedFilters,
@@ -930,13 +945,6 @@ export function BuildFilters({
                   <BaseListbox
                     className="mt-1"
                     value={unappliedFilters.withReference}
-                    onBlur={() => {
-                      if (
-                        unappliedFilters.withReference !== filters.withReference
-                      ) {
-                        applyUrlFilters(unappliedFilters);
-                      }
-                    }}
                     onChange={(value) => {
                       const newFilters = {
                         ...unappliedFilters,
@@ -962,14 +970,6 @@ export function BuildFilters({
                   <BaseListbox
                     className="mt-1"
                     value={unappliedFilters.withPatchAffected}
-                    onBlur={() => {
-                      if (
-                        unappliedFilters.withPatchAffected !==
-                        filters.withPatchAffected
-                      ) {
-                        applyUrlFilters(unappliedFilters);
-                      }
-                    }}
                     onChange={(value) => {
                       const newFilters = {
                         ...unappliedFilters,