+++ apps/remnant2toolkit/app/(builds)/_features/filters/_libs/filters/fusion-filter.ts
@@ -1,8 +1,7 @@
 import type { FilterOption } from '@repo/ui';
 
+import type { FilterDefinition } from '@/app/(builds)/_features/filters/_types/filter-definition';
 import { fusionItems } from '@/app/(items)/_constants/fusion-items';
-
-import { type FilterDefinition } from '../_types/filter-definition';
 
 export const fusionFilter = {
   buildFilterKey: 'fusions',